Add rabbitMQ channel pools to service aggregator
